Implementado RF008 (Consultar Questão) e RF014 (Responder Questão)
Primeira versão do realce de sintaxe dos editores de texto

+++ src/java/persistence/pergunta/PerguntaDAO.java
@@ -15,7 +15,7 @@ public interface PerguntaDAO
 {
     public void cadastrarPergunta(Pergunta p); //CREATE
     
-    public Pergunta getPergunta(Long id); //READ
+    public Pergunta getPergunta(Integer id); //READ
     
     public void atualizarPergunta(Pergunta p); //UPDATE
